requirements.txt
instructor==1.7.2
autogen==0.7.3
openai==1.58.0
ipdb
pyyaml
numpy>=1.21.0
scipy>=1.7.0
pandas>=1.3.0
scikit-learn==1.6.1
pydantic==2.10.6
accelerate==1.3.0
matplotlib
seaborn
transformers
langgraph>=0.0.23
langchain>=0.0.267
langchain-openai>=0.0.2.post1
networkx>=3.1
typing-extensions>=4.5.0